+++ server/src/lexer/__tests__/lexer.test.ts
@@ -213,6 +213,36 @@ describe('Lexer', () => {
     it('scans @env12:', () => {
       expect(tokenTexts('@env12:')).toEqual(['@env12:']);
     });
+
+    it('scans @env255:', () => {
+      expect(tokenTypes('@env255:')).toEqual([TokenType.EnvSpecifier]);
+      expect(tokenTexts('@env255:')).toEqual(['@env255:']);
+    });
+
+    it('scans @env0: prefixed in front of a keyword selector', () => {
+      // @env0:show: should produce two tokens: EnvSpecifier then Keyword
+      expect(tokenTypes('@env0:show:')).toEqual([
+        TokenType.EnvSpecifier,
+        TokenType.Keyword,
+      ]);
+      expect(tokenTexts('@env0:show:')).toEqual(['@env0:', 'show:']);
+    });
+
+    it('scans @env1: prefixed in front of a binary selector', () => {
+      expect(tokenTypes('@env1:+')).toEqual([
+        TokenType.EnvSpecifier,
+        TokenType.BinarySelector,
+      ]);
+      expect(tokenTexts('@env1:+')).toEqual(['@env1:', '+']);
+    });
+
+    it('scans @env2: prefixed in front of an identifier (unary)', () => {
+      expect(tokenTypes('@env2:squared')).toEqual([
+        TokenType.EnvSpecifier,
+        TokenType.Identifier,
+      ]);
+      expect(tokenTexts('@env2:squared')).toEqual(['@env2:', 'squared']);
+    });
   });
 
   describe('complex examples', () => {